Using the right types.

+++ AOFlagger.py
@@ -13,7 +13,7 @@ def test_swap():
     input_size = 8192
     array_device = numpy.random.randn(input_size).astype(numpy.float32)
     # array_host = array_device
-    arguments = [array_device, 0, 8191]
+    arguments = [array_device, numpy.uint32(0), numpy.uint32(8191)]
     parameters = []
     kernel_tuner.run_kernel("swap", kernel_source, input_size, arguments, parameters)
